% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_of_thrones.R
\docType{data}
\name{game_of_thrones}
\alias{game_of_thrones}
\title{Game of Thrones Data}
\format{
Nodes: 107; unimodal
Edges: 353; weighted; undirected

  This dataset is from Melanie Walsh. Prepared for R by Benjamin Smith
}
\source{
GitHub,
  <https://github.com/melaniewalsh/sample-social-network-datasets>
}
\usage{
game_of_thrones
}
\description{
Network data for character relationships within George R. R. Martin's \emph{A Storm of Swords}, the third novel in his series \emph{A Song of Ice and Fire} (also known as the HBO television adaptation \emph{Game of Thrones}).
}
\details{
This data was originally compiled by A. Beveridge and J. Shan, "Network of Thrones", Math Horizons Magazine , Vol. 23, No. 4 (2016), pp. 18-22.


This Game of Thrones data lends itself well to basic social network analysis because there are a lot of characters (and yet not \emph{too many} characters) and because, as the authors claim, these characters are "scattered geographically and enmeshed in their own social circles," such as various noble houses.

For those unacquainted with GOT, here's a bit more background info (via Beveridge and Shan) that might be helpful for understanding the social network:

 "\emph{The narrative starts at a time of peace, with all the houses unified under the rule of King Robert Baratheon, who holds the Iron Throne...[Then King Robert dies and all heck breaks loose.]}

\emph{Driven by cause or circumstance, characters from the many noble families launch into arduous and intertwined journeys. Among these houses are the honorable Stark family (Eddard, Catelyn, Robb, Sansa, Arya, Bran, and Jon Snow), the pompous Lannisters (Tywin, Jaime, Cersei, Tyrion, and Joffrey), the slighted Baratheons (led by Robert’s brother Stannis) and the exiled Daenerys, the last of the once powerful House Targaryen.}"

If you'd still like to know more about the source novel, you might also check out the \href{https://en.wikipedia.org/wiki/A_Storm_of_Swords}{Wikipedia page} or \href{https://www.youtube.com/watch?v=6N4gEJ_ED98}{this cheeky 8-minute beginner's guide} to the HBO show (warning: there are curse words aplenty).
}
\keyword{datasets}
