% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphwalshdata-package.R
\docType{package}
\name{igraphwalshdata-package}
\alias{igraphwalshdata}
\alias{igraphwalshdata-package}
\title{igraphwalshdata: 'igraph' Datasets from Melanie Walsh}
\description{
Interesting 'igraph' datasets from Melanie Walsh's sample social network datasets repository \url{https://github.com/melaniewalsh/sample-social-network-datasets}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benyamindsmith/igraphwalshdata}
  \item Report bugs at \url{https://github.com/benyamindsmith/igraphwalshdata/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Smith \email{benyamin.smith@mail.utoronto.ca} (\href{https://orcid.org/0009-0007-2206-0177}{ORCID})

}
\keyword{internal}
