% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political_books_network.R
\docType{data}
\name{political_books_network}
\alias{political_books_network}
\title{Political Books Social Network Data}
\format{
Nodes: 105; unimodal
Edges: 441; unweighted; undirected
  This dataset is from Melanie Walsh. Prepared for R by Benjamin Smith
}
\source{
GitHub,
  <https://github.com/melaniewalsh/sample-social-network-datasets>
}
\usage{
political_books_network
}
\description{
Krebs selected political books from the *New York Times* bestseller list, searched for those books within Amazon and Barnes&Noble, and then recorded a co-purchasing relationship between the two books if they were connected by a 'customers who bought this book also bought...' link.
}
\details{
For more on the methodology, see [Krebs's website](http://www.orgnet.com/divided2.html) or [the New York Times feature](http://www.nytimes.com/2004/03/13/books/study-finds-a-nation-of-polarized-readers.html) on his work.

Though this network is not technically, or perhaps not exclusively, a social one, the network reveals and suggets potential patterns about the communities of people buying these books. It also calls into question how forms of media consumption might influence, shape, reinforce, or reflect national political polarization.
}
\keyword{datasets}
